requirements.txt
streamlit==1.31.1
google-generativeai==0.3.2
python-dotenv==1.0.1
altair==5.2.0
blinker==1.7.0
cachetools==5.3.2
click==8.1.7
importlib-metadata==7.0.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.0
requests==2.31.0
rich==13.7.0
tenacity==8.2.3
toml==0.10.2
typing-extensions==4.9.0
tzlocal==5.2
validators==0.22.0
gitpython==3.1.42
pydeck==0.8.1b0
watchdog==4.0.0
google-ai-generativelanguage==0.4.0
google-auth==2.28.0
google-api-core==2.17.1
tqdm==4.66.2
proto-plus==1.23.0
jinja2==3.1.3
jsonschema==4.21.1
toolz==0.12.1
colorama==0.4.6
gitdb==4.0.11
googleapis-common-protos==1.62.0
pyasn1-modules==0.3.0
rsa==4.9
zipp==3.17.0
pytz==2024.1
tzdata==2024.1
charset-normalizer==3.3.2
idna==3.6
urllib3==2.2.0
certifi==2024.2.2
markdown-it-py==3.0.0
smmap==5.0.1
grpcio==1.60.1
grpcio-status==1.60.1
MarkupSafe==2.1.5
attrs==23.2.0
jsonschema-specifications==2023.12.1
referencing==0.33.0
rpds-py==0.18.0
mdurl==0.1.2
pyasn1==0.5.1
markdown-it-py

#invoice
#streamlit
#google-generativeai
#python-dotenv
#langchain
#PyPDF2 (load pdf or read any pdf)
#chromadb (vector embeddings) '''

langchain
PyPDF2
chromadb


#Reading multiple PDFs

#streamlit
#google-generativeai
#python-dotenv(load env variables)
#langchain(to read the pdf)
#PyPDF2
#faiss-cpu / more files can use files-gpu
#langchain_google_genai (with the help of langchain will be able to access the api of google-genai



faiss-cpu
langchain_google_genai

#youtube_transcript_api
#streamlit
#google-generativeai
#python-dotenv
#pathlib

youtube_transcript_api
pathlib
